Fixed the symmetric splitter.
We still have undesirable behaviour in the
`ReplaceContent#validate_endpoints` method, where an out-of-range
*beginning* endpoint will trigger a `RangeError` and be caught and
reported properly, but an out-of-range *ending* endpoint is treated as
equivalent to the length of the content or to -1; i.e., the end of the
content string. Implementing an endpoints-end error check and treating
an out-of-bounds error as invalid would require more code than the issue
appears to warrant, as we are easily convinced that "to infinity and
beyond" should be functionally equivalent to "to the end of the string".

Hence, we are [Closing #13].

46 tests, 51 assertions, 0 failures, 0 errors, 0 skips
Coverage: 412 / 412 LOC (100.0%) covered
RuboCop: 12 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 138.7; method average 3.1; max 5.9 (Prolog::Services::ReplaceContent::Splitter::Factory::Internals::_build_splitter_marker_params)
Reek: 0 warnings